requirements.txt
# streamlit==1.32.2
google-api-python-client==2.114.0
langchain==0.3.2
langchain-community==0.3.1
langchain-core==0.3.8
langchain-text-splitters==0.3.0
langchain-google-genai==2.0.0
datasets==3.0.1
scikit-learn==1.5.2
seaborn==0.13.2
umap-learn==0.5.6
numpy<2.0
